使用对象通用池来管理Cipher对象，优化性能同时避免线程安全
