+++ build.gradle
@@ -66,6 +66,7 @@ publishing {
 
 dependencies {
     implementation group: 'org.bouncycastle', name: 'bcpkix-jdk15on', version: '1.67'
+    implementation group: 'org.apache.commons', name: 'commons-pool2', version: '2.9.0'
     testCompile group: 'junit', name: 'junit', version: '4.12'
     testCompile group: 'org.apache.commons', name: 'commons-lang3', version: '3.11'
 }
