+++ src/main/java/twgc/gm/sm2/SM2Util.java
@@ -1,3 +1,5 @@
+package twgc.gm.sm2;
+
 import java.io.File;
 import java.io.FileReader;
 import java.io.IOException;
