+++ src/test/java/SM2UtilThreadTest.java
@@ -8,6 +8,7 @@ import org.junit.Assert;
 import org.junit.FixMethodOrder;
 import org.junit.Test;
 import org.junit.runners.MethodSorters;
+import twgc.gm.sm2.SM2Util;
 
 
 /**
