+++ src/test/java/SM3UtilThreadTest.java
@@ -4,7 +4,7 @@ import java.util.concurrent.ConcurrentLinkedQueue;
 import org.apache.commons.lang3.RandomStringUtils;
 import org.junit.Assert;
 import org.junit.Test;
-
+import twgc.gm.sm3.SM3Util;
 
 
 /**
