+++ src/test/java/SM4UtilTest.java
@@ -1,14 +1,10 @@
 import java.nio.charset.Charset;
-import java.security.InvalidAlgorithmParameterException;
-import java.security.InvalidKeyException;
 import java.security.NoSuchAlgorithmException;
 import java.security.NoSuchProviderException;
 import java.util.Arrays;
 import java.util.Collection;
 import java.util.Queue;
 import java.util.concurrent.ConcurrentLinkedQueue;
-import javax.crypto.BadPaddingException;
-import javax.crypto.IllegalBlockSizeException;
 import javax.crypto.NoSuchPaddingException;
 
 import org.apache.commons.lang3.RandomStringUtils;
@@ -17,6 +13,8 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.Parameterized;
 import org.junit.runners.Parameterized.Parameters;
+import twgc.gm.sm4.SM4ModeAndPaddingEnum;
+import twgc.gm.sm4.SM4Util;
 
 /**
  * @author Sean
@@ -30,25 +28,25 @@ public class SM4UtilTest {
     private static byte[] content16 = new byte[]{1, 2, 3, 4, 5, 6, 7, 8, 1, 2, 3, 4, 5, 6, 7, 8};
     //private static byte[] key = null;
     private static byte[] iv = null;
-    private static SM4Util.SM4ModeAndPaddingEnum type;
+    private static SM4ModeAndPaddingEnum type;
     static int randomData = 128;
     static String message = RandomStringUtils.random(randomData);
     static String exceptionHappened = "Exception happened";
 
     @Parameters(name = "{index}: sm4({1})")
-    public static Collection prepareData() throws NoSuchProviderException, NoSuchAlgorithmException {
+    public static Collection prepareData() {
         Object[][] object = {
-                {content16, SM4Util.SM4ModeAndPaddingEnum.SM4_ECB_NoPadding, false},
-                {message.getBytes(Charset.forName("utf8")), SM4Util.SM4ModeAndPaddingEnum.SM4_ECB_PKCS5Padding, false},
-                {message.getBytes(Charset.forName("utf8")), SM4Util.SM4ModeAndPaddingEnum.SM4_ECB_PKCS7Padding, false},
-                {content16, SM4Util.SM4ModeAndPaddingEnum.SM4_CBC_NoPadding, true},
-                {message.getBytes(Charset.forName("utf8")), SM4Util.SM4ModeAndPaddingEnum.SM4_CBC_PKCS5Padding, true},
-                {message.getBytes(Charset.forName("utf8")), SM4Util.SM4ModeAndPaddingEnum.SM4_CBC_PKCS7Padding, true}
+                {content16, SM4ModeAndPaddingEnum.SM4_ECB_NoPadding, false},
+                {message.getBytes(Charset.forName("utf8")), SM4ModeAndPaddingEnum.SM4_ECB_PKCS5Padding, false},
+                {message.getBytes(Charset.forName("utf8")), SM4ModeAndPaddingEnum.SM4_ECB_PKCS7Padding, false},
+                {content16, SM4ModeAndPaddingEnum.SM4_CBC_NoPadding, true},
+                {message.getBytes(Charset.forName("utf8")), SM4ModeAndPaddingEnum.SM4_CBC_PKCS5Padding, true},
+                {message.getBytes(Charset.forName("utf8")), SM4ModeAndPaddingEnum.SM4_CBC_PKCS7Padding, true}
         };
         return Arrays.asList(object);
     }
 
-    public SM4UtilTest(byte[] content, SM4Util.SM4ModeAndPaddingEnum type, boolean flag) throws NoSuchProviderException, NoSuchAlgorithmException, NoSuchPaddingException {
+    public SM4UtilTest(byte[] content, SM4ModeAndPaddingEnum type, boolean flag) throws NoSuchProviderException, NoSuchAlgorithmException, NoSuchPaddingException {
         this.content = content;
         this.type = type;
         if (flag) {
